requirements.txt
requests
beautifulsoup4
langchain
langchain-community
langchain-core
pypdf
langchain-openai
python-dotenv
faiss-cpu
rapidocr-onnxruntime
